+++ packages/bylaw_credo/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- Add `Bylaw.Credo.Check.Ecto.PreferOrderByOverRepoAllEnumSort` to discourage
+  sorting `Repo.all` results in memory instead of using `order_by`.
+
 ## 0.3.0 - 2026-08-06
 
 - Add `Bylaw.Credo.Check.HEEx.DesignSystem.NoArbitrarySpacing` to discourage